patch: More verbose yaml errors
When the docstring is invalid yaml/swagger, the generated swagged documentation
just shows:

```
Invalid Swagger
⚠ Swagger document could not be loaded from docstring ⚠
```

This change makes the docstring parsing step a bit more verbose to indicate that
there are problems. It is still not ideal but it is an improvement over guessing/
copy-pasta into a validator.

How to test that this is working:
1) create a aiohttp route with docstring, add it to the app with the swagger plugin
2) start the app
3) navigate to the /api/doc page and find the route
   observe that the route is displayed correctly in the swagger page
4) change the docstring to invalid yaml (e.g. by indenting with tabs)
5) (re) start the app
   observe that there is more verbose output
6) (re) navigat ot the /api/doc page and find the route
   observe that the route now probably is labeled as 'Invalid Swagger'